feat(iframe)!: remove `initialState`
BREAKING CHANGE: Setting an initial state for a cookie (i.e. enabled instead of disabled) was never properly implemented. To avoid confusion, this setting is removed. It continues to be possible to set `isIframeBlocked` true, though as of 2022-03-30 it is not yet implemented for vite (https://github.com/dargmuesli/nuxt-cookie-control/issues/79).