feat(module)!: require cookie id
When no cookie id was given, the cookie id was set to the slugified version of the cookie name. This
lead to a large bundle size and made it less obvious that a short cookie id would save bandwidth.
Therefore the slugification along with the `getCookieId` method is removed and it is now required to
set cookie ids manually, improving the project's performance.